Add a new SDHCI quirk, SDHCI_QUIRK_DONT_SET_HISPD_BIT.  Apparently some sdhci controllers, such as the one on a Raspberry Pi, mishandle the signal timing in high speed signaling mode, but run just fine in standard mode with the bus running at frequencies between 25-50MHz (which shouldn't work).
This is the solution adopted by U-Boot and other OSes (linux and *BSD)
for the timeouts on Raspberry Pi boards with certain SD cards.  Some
research shows that this quirk is also used on a few other boards, so the
fix is a generic quirk instead of being in the RPi-specific driver code.

This change is based on information discovered by Michal Meloun.


git-svn-id: https://svn.freebsd.org/base/head@277307 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f