Add a CLOUDPACKAGE variable to contain a list of targets for post-install packaging cloud provider images.
Add a 'gce-package.sh' script to generate the final output
image ready for upload to the GCE platform.  Right now, this
is the only image that has a specific output format (GNU-tar),
and this implementation is expected to be temporary.

This is not directly connected to the other release targets.

MFC after:	1 week
X-MFC-with:	r279249, r279250
Sponsored by:	The FreeBSD Foundation


git-svn-id: https://svn.freebsd.org/base/head@279257 ccf9f872-aa2e-dd11-9fc8-001c23d0bc1f